Add add workflow hook examples for build-and-push workflow (#32)
In this commit, we're starting to experiment with what kinds of patterns we want to suggest or standardize as a part of the paved path for building and publishing Docker containers. These are the aspects that we'll want to highlight and provide documentation for in Confluence:

- adding a HEALTHCHECK instruction to the Dockerfile

- adding a Makefile with these targets that we expect the build-and-push workflow to call:
    - make mozcloud-prebuild
    - make mozcloud-postbuild

- the kinds of checks and tasks to run in these {pre,post}build commands
  
This commit uses Docker Compose as the way we build containers and run tests. This is because many teams use both Make and Docker Compose to orchestrate local development tasks like building and running tests. Though we'll still want teams to eventually use our actions since they'll offer an easy way to build containers the right way (proper tagging, future improvements like build attestations), we still want to offer suggestions on how to make any custom workflows the most compatible with our actions and reusable CI workflows.